// network/types.ts
export interface User {
  email: string;
  profile?: {
    first_name?: string;
    last_name?: string;
    phone_number?: string;
    address?: string;
  } | null;
}